requirements.txt
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.5.0
inference==0.9.8  # from roboflow_model.py (InferenceHTTPClient)
torchvision>=0.15.0
